+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `BytesFile` adapter for `TableStoreFiledir`
 * Added `delete_stale` argument to `BatchGenerate`
+* Fix duplicated indices in `BaseBatchTransformStep.get_full_process_ids`
 
 # 0.13.6